+++ NEWS.md
@@ -17,7 +17,9 @@
 * Added [VPdtw](https://github.com/ethanbass/VPdtw/) as a dependency (instead of being only suggested).
 * Minor updates to vignette.
 * Adjusted `cluster_spectra` and `combine_peaks` functions so messages can be suppressed with `verbose == FALSE`.
-* Fixed test failure on MKL server by skipping `cluster_spectra` test on CRAN.
+* Fixed occasional test failure on MKL server by skipping `cluster_spectra` test on CRAN.
+* Updated find_peaks function with more and better smoothing options. The defaults smoothing is slightly changed.
+Find peaks will now do gaussian smoothing by default with
 
 #### Changes to *fit_peaks* function: